Add different item list

+++ app/src/main/java/com/shkiper/taskmanager/viewmodels/MainViewModel.kt
@@ -1,6 +1,9 @@
 package com.shkiper.taskmanager.viewmodels
 
+import android.view.View
+import android.widget.TextView
 import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
 import com.shkiper.taskmanager.models.Task
@@ -11,8 +14,11 @@ class MainViewModel : ViewModel() {
     private val taskRepository = TaskRepository
     private val tasks = Transformations.map(taskRepository.loadChats()) { tasks ->
         return@map tasks.filter {!it.isDone}
-            .sortedBy { it.id.toInt() }
+            .sortedBy { it.id }
     }
+
+
+
 
     fun getTaskData(): LiveData<List<Task>> {
         return tasks
